WIP: add prefix scan methods to do e.g. cumsum and cumprod.
Added:
- `my_matrix.ss.scan_rows(optional_monoid)`
- `my_matrix.ss.scan_columns(optional_monoid)`
- `my_vector.ss.scan(optional_monoid)`

We perform the scan with a series of matrix multiplies a la the Blelloch method.  See:
https://developer.nvidia.com/gpugems/gpugems3/part-vi-gpu-computing/chapter-39-parallel-prefix-sum-scan-cuda

I bet there are faster ways to do this and that the current approach could be further optimized.
Nevertheless, I think this is a nice and useful API, so I'd like to add it to help push the boundary.

We could turn this into an expression so that masks, accumulators, etc. could be used,
but I haven't worried about such semantics yet (and don't know of a use case).
We could also further extend this perform a transpose, such as `my_matrix.T.ss.scan_rows()`.

TODO:
- [] document
- [] test use of different monoids
- [] improve and test `get_semiring`
- [] benchmark and optimize?