Make `v += expr` do `v(plus) << expr` when expr is not scalar.
This allows `+=` (or any infix that performs outer join) do operate in one
C call instead of two.

Pragmatically, this helps one from debating over `x += expr` and `x(plus) << expr`.
This keeps the original behavior with scalars: `x += 1` is the same as `x << x + 1`,
which only adds 1 to existing entries.  It does not "densify" the object.