Assign list to many indices, such as `v[[1, 2]] = [3, 4]`
Closes #245.  This also fixes some import functions for sub-array UDTs.
Shapes must match exactly.  We aren't forgiving or do broadcasing like NumPy.

I agree in principle that *if* we were to allow operations between
e.g. vectors and numpy arrays or lists (such as `binary.min(v & a)`),
then we should probably be consistent with how we treat
lists/arrays/etc as vectors or matrices when assigning.
I think we should treat them as dense in all cases.

I don't think using a list to assign multiple items at a time *implies*
that a list or array should be treated as vectors or matrices elsewhere.
The mental model is that it's assigning multiple elements.  That's it.
That we convert it to a Vector or Matrix is an implementation detail.

I think assigning lists/arrays is intuitive, and useful in an interactive,
educational setting.  It's plausible that it will be useful in practice too
to conveniently and efficiently assign arrays to sub-vector/sub-matrices.

Adding this functionality was pretty straightforward and very isolated.
Most of the code is to enable good error messages and handling sub-array UDTs.